FIX: update sys.path depths after scripts/ reorganization; fix test imports
