+++ scripts/training/train_hybrid.py
@@ -32,7 +32,7 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 
-sys.path.insert(0, str(Path(__file__).parent.parent))
+sys.path.insert(0, str(Path(__file__).parent.parent.parent))
 
 from sklearn.metrics import f1_score, precision_score, recall_score
 
@@ -55,7 +55,7 @@ from src.traditional_methods import (
     select_best_starts,
 )
 
-MODELS_DIR = Path(__file__).parent.parent / "models"
+MODELS_DIR = Path(__file__).parent.parent.parent / "models"
 DATA_DIR = get_data_dir("full_dataset")
 PROD_MODEL = MODELS_DIR / "hybrid_best_model.pkl"
 NEW_MODEL = MODELS_DIR / "hybrid_best_model_v2.pkl"
